alt-theory-rag/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "knowledge-rag"
version = "3.2.3"
description = "Local RAG System for Claude Code — Hybrid search + Cross-encoder Reranking + 12 MCP Tools. Zero external servers."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
authors = [
    {name = "Lyon.", email = "lyonzin@users.noreply.github.com"},
]
keywords = [
    "rag", "mcp", "claude-code", "semantic-search", "embeddings",
    "chromadb", "bm25", "reranking", "knowledge-base", "local-ai",
    "fastembed", "retrieval-augmented-generation", "hybrid-search",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
]
dependencies = [
    "chromadb>=1.4.0",
    "pymupdf>=1.23.0",
    "fastembed[reranking]>=0.4.0",
    "mcp>=1.0.0",
    "rank-bm25>=0.2.2",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=1.0.0",
    "watchdog>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/lyonzin/knowledge-rag"
Repository = "https://github.com/lyonzin/knowledge-rag"
Issues = "https://github.com/lyonzin/knowledge-rag/issues"
Changelog = "https://github.com/lyonzin/knowledge-rag/releases"

[project.scripts]
knowledge-rag = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[tool.hatch.build.targets.sdist]
include = [
    "mcp_server/",
    "documents/examples/",
    "README.md",
    "LICENSE",
    "pyproject.toml",
    "requirements.txt",
]
exclude = [
    "data/",
    "documents/security/",
    "documents/ctf/",
    "documents/logscale/",
    "documents/development/",
    "documents/general/",
    "documents/aar/",
    "venv/",
    "__pycache__/",
    "*.pyc",
    "scripts/",
    ".gitignore",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line length handled by formatter

[tool.coverage.run]
source = ["mcp_server"]

[tool.coverage.report]
show_missing = true
fail_under = 35
